Fix accidental use of uninitialized memoroy.
In the original private project, the new operator had a custom
implementation that automatically zero-filled the allocated memory.

That's been intentionally removed, but that exposed a quirk in the
OptionsTemplate<> where it looped until m_max instead of m_count, and
with the removal of the non-standard new operator, that made it refer
to uninitialized entries in the list of parsed options.