Refactor state management due to scan vs render phases.
The scan phase calculates all the statistics and widths.
The render phase renders using the calculations from the scan phase.

But the field widths were still being tracked essentially in a
singleton, which meant the render phase used the same state for all
directories.

There are several ways to handle that, and I chose the brute force
approach of having a separate PictureFormatter instance per DirContext.

This is the most thorough and requires the least maintenance over the
long term.  Since generally speaking the number of directories is
eclipsed by the number of files, the performance cost of instantiating a
new PictureFormatter for each directory should be negligible.

Especially since the PictureFormatter is relatively lightweight anyway.
It contains two vectors, and they're proportional to the number of
fields.  The number of fields is always small, except in the pathological
case where a custom format picture shows many fields redundantly.