Fix how --tree tracks directories.
Previously using a wildcard accidentally suppressed recursing into
directory names that didn't match the wildcard.  The intent is to
recurse into subdirectories regardless of their names, and apply the
wildcard to the files in the directory.

This is ambiguous either way, though, in `--tree` mode:
- Applying wildcard to dir names would prevent seeing files in those
  directories.
- Not applying wildcard to dir names feels inconsistent compared to how
  `dir` and `dirx` normally operate.

But the best compromise in order for `--tree` to be able to show a tree
view is to list directory names even if they don't match the
wildcard(s).

This ends up showing the entire directory structure when there are no
matching files.  An alternative could be to post-process the lists and
discard directories that have no matching files.  But that gets weird
and inconsistent, because `--tree` without a wildcard needs to list
empty directories.

The behavior might change in the future, but for now this seems like a
functional compromise.